requirements.txt
openai>=1.0.0
anthropic>=0.3.0
pyyaml>=6.0
python-dotenv>=1.0.0
tenacity>=8.2.0
tiktoken>=0.5.0
pydantic>=2.0.0

